requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
timm>=0.9.0
segmentation-models-pytorch>=0.3.0
albumentations>=1.3.0
opencv-python>=4.7.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.2.0
pycocotools>=2.0.6
tqdm>=4.65.0
tensorboard>=2.13.0
numpy>=1.24.0
pillow>=9.5.0
pyyaml>=6.0
